pyproject.toml
[project]
name = "planifia"
version = "0.1.0"
description = "A custom made project hours management"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.7",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2025.3.1",
    "pillow>=11.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-django>=4.10.0",
    "ruff>=0.9.10",
]
